+++ libraries/botbuilder-adapter-alexa/src/alexaSchema.ts
@@ -1,0 +1,5 @@
+export enum AlexaActivityTypes {
+    LaunchRequest = 'LaunchRequest',
+    SessionEndedRequest = 'SessionEndedRequest',
+    IntentRequest = 'IntentRequest'
+}