+++ libraries/botbuilder-adapter-twitter/CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2019-10
-### Added
-- Initial release for Twitter Adapter
+## [1.0.2] - 2020-03
+### Changed
+- Updated dependencies
 
 ## [1.0.1] - 2019-11
 ### Fixed
 - Moved Twitter package types to `dependencies` from `devDependencies`
+
+## [1.0.0] - 2019-10
+### Added
+- Initial release for Twitter Adapter